Vil for statistikk-uthentinga av behandling ha behov for å kunne hente ut behandling. Skil derfor ut den delen av behandlingservice som er henting til si eiga klasse. Tenkjer uansett det er fint å få ned mengden samtidige ting BehandlingService har ansvaret for
